Moved socket methods to a mixin.
