+++ src/lib/y2remote/modes/vnc.rb
@@ -1,47 +1,24 @@
 require "y2remote/modes/base"
+require "y2remote/modes/sockets_mixin"
 
 module Y2Remote
-  class Modes
+  module Modes
     class VNC < Base
+      include SocketsMixin
+
       SOCKET = "xvnc".freeze
+
+      def initialize
+        super()
+        Yast.import "Packages"
+      end
 
       def required_packages
         Yast::Packages.vnc_packages
       end
 
-      def socket
-        Yast::SystemdSocket.find(SOCKET)
-      end
-
-      def enabled?
-        return false unless socket
-
-        socket.enabled?
-      end
-
-      def enable!
-        return false unless socket
-
-        socket.enable
-      end
-
-      def disable!
-        return false unless socket
-
-        socket.disable
-      end
-
-      def stop!
-        return false unless socket
-
-        socket.stop
-      end
-
-      def restart!
-        return false unless socket
-
-        stop! if socket.listening?
-        socket.start
+      def socket_name
+        SOCKET
       end
     end
   end
